feat: add kubectl_exec to kubernetes tools, remove network-diagnostics.ts
kubectl_exec (namespace, pod, command, optional container) covers pod
connectivity testing, DNS resolution, and any other in-pod debugging —
making the bespoke netdiag_* wrappers redundant.
